requirements-proxy.txt
flask
flask-cors
requests
python-dotenv
huggingface_hub
pillow
